requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
zipp==3.7.0
